demo ng-model directive
on an input box. now you can type in the box and angular will
automatically populate the object you name will the value from the box

this is one of the coolest features of angular. You dont have to
specifically wire angular to the DOM yourself or write your own event
binding here